apps/sam3-backend/pyproject.toml
[project]
name = "sam3-backend"
version = "0.10.0"
description = "SAM 3 ML Backend (facebookresearch/sam3) for ai-annotation-platform"
requires-python = ">=3.12,<3.13"

# torch / torchvision 由 docker base image (pytorch/pytorch:2.7.0-cuda12.6-cudnn-devel) 锁定,
# 不在此处重复声明以避免 pip 误升级覆盖预装版本.
# 与官仓 facebookresearch/sam3 的 README 对齐: CUDA 12.6+, Python 3.12, torch 2.7+.
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "httpx>=0.27",
    "numpy>=1.26,<3.0",
    "pillow>=10.0",
    "opencv-python-headless>=4.9",
    "shapely>=2.0",
    # huggingface_hub: SAM 3.1 是 gated repo, 拉权重必须传 HF_TOKEN.
    "huggingface_hub>=0.23",
    # transformers 5.x 对 SAM 3 友好 (Sam3Model / Sam3VideoModel 在 5.x 落地).
    # 具体下限以 facebookresearch/sam3 README 为准, sync_vendor.sh 升级后核对.
    "transformers>=4.45",
    "supervision>=0.21",
    "prometheus-client>=0.20",
    # PerfHud · GPU util/温度/功耗 + 容器 CPU/RAM
    "pynvml>=11.5",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "predictor", "schemas", "embedding_cache", "observability"]

[tool.pytest.ini_options]
testpaths = ["tests"]
# mask_utils 共享包从 apps/_shared/mask_utils 注入. Docker 容器内由
# `pip install -e /app/mask_utils` 提供; 本地 pytest 通过 pythonpath 直接定位 src/.
pythonpath = [".", "../_shared/mask_utils/src"]
